FBReader 0.12.0

+++ fbreader/src/database/booksdb/runnables/FindFileIdRunnable.cpp
@@ -42,7 +42,7 @@ bool FindFileIdRunnable::run() {
 	DBIntValue &addParent = (DBIntValue &) *myAddFile->parameter("@parent_id").value();
 	((DBIntValue &) *myAddFile->parameter("@size").value()) = 0;
 
-	int index = dirName.length() + 1;
+	size_t index = dirName.length() + 1;
 	findName = dirName;
 	findParent = 0;
 	while (true) {
@@ -62,7 +62,7 @@ bool FindFileIdRunnable::run() {
 			myFileId = ((DBIntValue &) *physId).value();
 			return true;
 		}
-		int index2 = resolvedPath.find(BooksDBQuery::ArchiveEntryDelimiter, index);
+		size_t index2 = resolvedPath.find(BooksDBQuery::ArchiveEntryDelimiter, index);
 		findName = resolvedPath.substr(index, index2 - index);
 		index = index2 + 1;
 		findParent = ((DBIntValue &) *physId).value();
